+++ history.rst
@@ -3,7 +3,7 @@ Google App Engine - Address Book - Data Backend
 ###############################################
 
 
-ToDo: Security: has_authorization(user_name, authorization_name)
+ToDo: Security: Authorization: *common.authorization*
 
 
 =============